+++ webpack.config.js
@@ -5,7 +5,7 @@ module.exports = {
   },
   devtool: 'inline-source-map',
   module: {
-    loaders: [
+    rules: [
       { test: /\.js$/, loader: 'babel-loader' },
     ],
   },