+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - update eslint packages
+- move to jest from mocha/chai
 
 ## [0.1.2] - 2017-09-08
 ### Fixed