+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - update eslint packages
 - move to jest from mocha/chai
 - update React and use PureComponent
+- update babel packages
 
 ## [0.1.2] - 2017-09-08
 ### Fixed