+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - move to jest from mocha/chai
 - update React and use PureComponent
 - update babel packages
+- update webpack packages
 
 ## [0.1.2] - 2017-09-08
 ### Fixed